+++ inst/tinytest/testPlot.R
@@ -952,6 +952,26 @@ expect_true_xl(
   inherits(json, "json")
 )
 
+# Catch plots of plotType "plotly" that don't correctly return a plotly plot
+notPlotlyFunc <- function(x) graphics::plot(1:10)
+notPlotlyFuncList <- list(
+  default = list(
+    notPlotlyFunc = list(
+      displayName = "notPlotly",
+      plotType = "plotly",
+      packages = "plotly"
+    )
+  )
+)
+
+testStudyNotPloty <- addPlots(testStudyObj, notPlotlyFuncList)
+expect_error(
+  plotStudy(testStudyNotPloty, modelID = "model_01", featureID = "feature_0001", plotID = "notPlotlyFunc"),
+  "The plotID \"notPlotlyFunc\" has plotType \"plotly\" but did not return an object with class \"plotly\"",
+  info = "Detect invalid plotly functions"
+)
+
+
 # Teardown ---------------------------------------------------------------------
 
 unloadNamespace(testPkgName)